feat: moved country flags to another module
