+++ ichub-frontend/src/features/part-discovery/components/submodel-addons/BaseAddon.tsx
