+++ ichub-frontend/src/features/part-discovery/components/submodel-addons/us-tariff-information/index.ts
